Rename SelectMenu2 -> SelectMenu
